Allow serialization of static vars via @PreserveStatic Allows configuration for serializtion to be generated with @SerializeOptions
Wrote tweaks to ConfigMode to have a few new methods of serializing

More emphasis on @Path

All unit Tested!
