+++ Yamler-Core/src/main/java/net/cubespace/Yamler/Config/ConfigMode.java
@@ -5,5 +5,6 @@ package net.cubespace.Yamler.Config;
  */
 public enum ConfigMode {
     DEFAULT,
+    PATH_BY_UNDERSCORE,
     FIELD_IS_KEY
 }
