+++ Yamler-Core/src/main/java/net/cubespace/Yamler/Config/PreserveStatic.java
@@ -1,0 +1,12 @@
+package net.cubespace.Yamler.Config;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface PreserveStatic {
+	boolean value() default true;
+}
