+++ Yamler-Core/src/main/java/net/cubespace/Yamler/Config/SerializeOptions.java
@@ -1,0 +1,16 @@
+package net.cubespace.Yamler.Config;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface SerializeOptions {
+	String[] configHeader() default {};
+
+	boolean skipFailedObjects() default false;
+
+	ConfigMode configMode() default ConfigMode.FIELD_IS_KEY;
+}
