+++ Yamler-Core/src/test/java/AnotherConfig.java
@@ -1,8 +1,8 @@
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class AnotherConfig extends Config {
+public class AnotherConfig extends YamlConfig {
     public String test = "test";
 }
