+++ Yamler-Core/src/test/java/FieldIsKeyConfig.java
@@ -1,10 +1,10 @@
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 import net.cubespace.Yamler.Config.ConfigMode;
 
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class FieldIsKeyConfig extends Config {
+public class FieldIsKeyConfig extends YamlConfig {
     public FieldIsKeyConfig() {
         CONFIG_MODE = ConfigMode.FIELD_IS_KEY;
     }
