+++ Yamler-Core/src/test/java/ListSubConfig.java
@@ -1,8 +1,8 @@
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class ListSubConfig extends Config {
+public class ListSubConfig extends YamlConfig {
     public String Test = "Test";
 }
