+++ Yamler-Core/src/test/java/SetConfig.java
@@ -1,11 +1,11 @@
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 import java.util.*;
 
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class SetConfig extends Config {
+public class SetConfig extends YamlConfig {
     public java.util.Set<String> StringSet = new HashSet<String>(){{
         add("Test");
     }};
