+++ Yamler-Core/src/test/java/converter/config/SimpleObjectConfig.java
@@ -1,6 +1,6 @@
 package converter.config;
 
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -9,7 +9,7 @@ import java.util.Set;
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class SimpleObjectConfig extends Config {
+public class SimpleObjectConfig extends YamlConfig {
     public java.util.Map TestMap = new HashMap<String, String>(){{
         put("test", "test");
     }};
