+++ Yamler-Core/src/test/java/performance/config/PerformanceConfig.java
@@ -2,7 +2,7 @@ package performance.config;
 
 import net.cubespace.Yamler.Config.Comment;
 import net.cubespace.Yamler.Config.Comments;
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 import net.cubespace.Yamler.Config.ConfigSection;
 
 import java.util.ArrayList;
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class PerformanceConfig extends Config {
+public class PerformanceConfig extends YamlConfig {
     @Comment("Database configuration.")
     private ConfigSection database;
     @Comments({"The database JDBC address. Should replace dbname with the database name."})
