+++ Yamler-Core/src/test/java/realconfig/gesuit/config/Announcements.java
@@ -1,7 +1,7 @@
 package realconfig.gesuit.config;
 
 
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 import realconfig.gesuit.config.sub.AnnouncementEntry;
 
 import java.util.HashMap;
@@ -9,7 +9,7 @@ import java.util.HashMap;
 /**
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
-public class Announcements extends Config {
+public class Announcements extends YamlConfig {
     public Boolean Enabled = true;
     public HashMap<String, AnnouncementEntry> Announcements = new HashMap<>();
 }
