+++ Yamler-Core/src/test/java/realconfig/noannounce/config/AnnouncerConfig.java
@@ -1,12 +1,12 @@
 package realconfig.noannounce.config;
 
 import net.cubespace.Yamler.Config.Comment;
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 
-public class AnnouncerConfig extends Config {
+public class AnnouncerConfig extends YamlConfig {
     @Comment("Order, either sequential or random")
     public String order = "sequential";
     @Comment("The server that should be asked for permissions (Most likely the hub server)")
@@ -18,14 +18,14 @@ public class AnnouncerConfig extends Config {
     public HashMap<String, MessageMap> servers = new HashMap<String, MessageMap>();
 
 
-    public static class MessageMap extends Config{
+    public static class MessageMap extends YamlConfig {
         public ArrayList<String> servers;
         public int offset;
         public int delay;
         public String permission;
         public ArrayList<Announcement> announcements = new ArrayList<Announcement>();
     }
-    public static class Announcement extends Config{
+    public static class Announcement extends YamlConfig {
         public String type;
         public String message;
 
@@ -39,7 +39,7 @@ public class AnnouncerConfig extends Config {
 
     public HashMap<String, BroadcastMap> nonannouncements = new HashMap<String, BroadcastMap>();
 
-    public static class BroadcastMap extends Config{
+    public static class BroadcastMap extends YamlConfig {
         public ArrayList<String> servers;
         public String permission;
         public Announcement announcement;
