+++ Yamler-Core/src/test/java/realconfig/servermenu/config/Menus.java
@@ -1,11 +1,11 @@
 package realconfig.servermenu.config;
 
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Menus extends Config {
+public class Menus extends YamlConfig {
     private String title;
     private List<Servers> servers = new ArrayList<>();
