+++ Yamler-Core/src/test/java/update/UpdateConfigConfig.java
@@ -4,12 +4,12 @@ package update;
  * @author geNAZt (fabian.fassbender42@googlemail.com)
  */
 
-import net.cubespace.Yamler.Config.Config;
+import net.cubespace.Yamler.Config.YamlConfig;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class UpdateConfigConfig extends Config {
+public class UpdateConfigConfig extends YamlConfig {
     public Boolean Enabled = false;
     public Map<String, Integer> restriction = new HashMap<String, Integer>(){{
         put( "bedwars.restriction.player", 1 );
